feat(delegate): popup dialog + list polish (D3+D5+D6+D7+D8)
- D3: convert DelegateEditView to dialog (pattern I.2 user edit)
- D5: add type icons to Receiver and Resource columns on the list
  (lookup normalizes lowercase enum values from the backend)
- D6+D8: fuse Type and Resource columns into one — the type becomes
  the row's leading icon
- D7: add Admin/Write checkbox tooltips with i18n texts recovered
  from the legacy plugin-id

Also: rename DelegateEditView → DelegateEditDialog, drop the now-unused
/id/delegate/new and /id/delegate/:id routes, align the actions column
width on the existing Groups/Users/Companies views (120 px, center).
